+++ app/artifact-qa-inspector/page.tsx
@@ -1,0 +1,5 @@
+import ArtifactQaInspector from "../../components/ArtifactQaInspector";
+
+export default function ArtifactQaInspectorPage() {
+  return <ArtifactQaInspector />;
+}